requirements.txt
# Natural Language Processing
langchain
langchain_community
pdfminer.six==20221105
InstructorEmbedding==1.0.1
sentence-transformers
huggingface-hub
transformers
protobuf==3.20.2; sys_platform != 'darwin'
protobuf==3.20.2; sys_platform == 'darwin' and platform_machine != 'arm64'
protobuf==3.20.3; sys_platform == 'darwin' and platform_machine == 'arm64'
auto-gptq==0.2.2
docx2txt
#unstructured==0.12.5
text-generation
urllib3
accelerate
bitsandbytes ; sys_platform != 'win32'
bitsandbytes-windows ; sys_platform == 'win32'
click
requests==2.31.0

redis==5.0.2
# gradio related
gradio
xmltodict==0.13.0
pypdf==4.0.2
lxml==4.9.4
python-dotenv==1.0.1
#protorpc
# Excel File Manipulation
#openpyxl
